+++ scripts/legacyMacros.js
@@ -459,6 +459,7 @@ export {spittingMimicAcid} from './macros/2014/monsterFeatures/monstrosity/spitt
 export {thunderousStrike} from './macros/2014/monsterFeatures/elemental/earthElementalMyrmidon/thunderousStrike.js';
 // Generic Monster Features
 export {abilityDrain} from './macros/2014/monsterFeatures/generic/abilityDrain.js';
+export {activityOnRest, activityOnRestActor} from './macros/2014/monsterFeatures/generic/activityOnRest.js';
 export {advantageDamageBonus} from './macros/2014/monsterFeatures/generic/advantageDamageBonus.js';
 export {auraDamageEnd} from './macros/2014/monsterFeatures/generic/auraDamage.js';
 export {autoGrapple} from './macros/2014/monsterFeatures/generic/autoGrapple.js';
@@ -479,6 +480,7 @@ export {packTactics} from './macros/2014/monsterFeatures/generic/packTactics.js'
 export {parry} from './macros/2014/monsterFeatures/generic/parry.js';
 export {reduce, reduceReduced} from './macros/2014/monsterFeatures/generic/reduce.js';
 export {reduceMaxHP} from './macros/2014/monsterFeatures/generic/reduceMaxHP.js';
+export {rerollSaveOnDamage, rerollSaveOnDamageActor} from './macros/2014/monsterFeatures/generic/rerollSaveOnDamage.js';
 export {spellTurning} from './macros/2014/monsterFeatures/generic/spellTurning.js';
 export {sunlightSensitivity} from './macros/2014/monsterFeatures/generic/sunlightSensitivity.js';
 export {surpriseAttack} from './macros/2014/monsterFeatures/generic/surpriseAttack.js';