+++ scripts/macros/2024/classFeatures/bard/fontOfInspiration.js
@@ -25,7 +25,13 @@ export let fontOfInspiration = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ]
 };