+++ scripts/macros/2024/classFeatures/cleric/light/improvedWardingFlare.js
@@ -19,7 +19,13 @@ export let improvedWardingFlare = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ]
 };