+++ scripts/macros/2024/classFeatures/cleric/light/radianceOfTheDawn.js
@@ -45,7 +45,13 @@ export let radianceOfTheDawn = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ]
 };