+++ scripts/macros/2024/classFeatures/cleric/war/guidedStrike.js
@@ -59,7 +59,13 @@ export let guidedStrike = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ],
     config: [