+++ CHANGELOG.md
@@ -1,6 +1,49 @@
+# 1.3.32 Release:
+## New Modern Content:
+- Sneak Attack
+- Cunning Strike
+- Devious Strikes
+## Update Notes:
+- Subclass automations that interact with sneak attack and cunning strike will be in a future update.
+
+# 1.3.31 Release:
+## New Modern Content:
+- War God's Blessing
+- Avatar of Battle
+## New Generic Monster Features:
+- Activity On Rest. Note: Example "GMF - Bite" item included in the "CPR Miscellaneous" compendium.
+- Added Remove on Save and Activities config options to Reduce Max HP.
+## Bug Fixes:
+- Chat Message Dark Mode
+## API Update Notes:
+- Rest triggers now includes `data` in the `trigger` object, which has a `longRest` property.
+  
+# 1.3.30 Release:
+## New Generic Monster Features:
+- Reroll Save On Damage
+## Bug Fixes:
+- Auto Push
+## Update Notes:
+- Added more animation options to detect magic.
+  
+# 1.3.29 Release:
+## Update Notes:
+- Added item and actor medkit events to the embedded macros editor.
+- Added tool check events to the embedded macros editor.
+
+# 1.3.28 Release:
+## Update Notes:
+- Added tool checks to the special durations section of the CPR Effect Medkit.
+## API Update Notes:
+- Added tool check events.
+
 # 1.3.27 Release:
+## New Modern Content:
+- War Priest
 ## Update Notes:
 - Added Chat Message Dark Mode setting.
+## API Update Notes:
+- Added item and actor medkit events.
 
 # 1.3.26 Release:
 ## Bug Fixes: